refactor(suite-native): send form moved to the parent screen component
Necessary for passing the submit button as a Screen component footer.
